feat: Add Docker setup documentation and testing script for PwnDoc-ng
- Introduced a comprehensive Docker README outlining the setup for both development and production environments, detailing service architecture and configuration.
- Updated docker-compose files to reflect new container names and configurations for development, including MongoDB, backend, frontend, and LanguageTool services.
- Added a testing script to validate Docker setups, ensuring services are running correctly and endpoints are accessible.
- Enhanced Nginx configurations for both development and production to support SSL and API proxying.
- Updated frontend and backend Dockerfiles to use the latest Node.js version and improve build processes.
