+++ docker-compose-dev.yml
@@ -3,7 +3,7 @@ services:
 
   mongodb:
     image: mongo:4.2.15
-    container_name: mongo-pwndoc-ng
+    container_name: mongo-pwndoc-ng-dev
     volumes:
       - mongo-data-dev:/data/db
     ports:
@@ -11,12 +11,14 @@ services:
     restart: always
     environment:
       - MONGO_DB:pwndoc
+    networks:
+      - backend
 
   backend:
     build:
       context: ./backend/
       dockerfile: Dockerfile.dev
-    container_name: pwndoc-ng-backend
+    container_name: pwndoc-ng-backend-dev
     entrypoint: ""
     working_dir: "/app"
     command: "npm run dev"
@@ -30,14 +32,23 @@ services:
       - ./backend/swagger.js:/app/swagger.js
     depends_on:
       - mongodb
+    environment:
+      - NODE_ENV=dev
+      - COLLAB_WEBSOCKET_PORT=8440
     restart: always
+    ports:
+      - 5252:5252
+      - 8440:8440
     links:
       - mongodb
+    networks:
+      - backend
 
   frontend-app:
     build:
       context: ./frontend/
       dockerfile: Dockerfile.dev
+    container_name: pwndoc-ng-frontend-dev
     volumes:
       - ./frontend/src:/app/src
       - ./frontend/ssl:/app/ssl
@@ -48,12 +59,21 @@ services:
       - ./frontend/tailwind.config.js:/app/tailwind.config.js
       - ./frontend/postcss.config.js:/app/postcss.config.js
       - ./frontend/tsconfig.json:/app/tsconfig.json
+      - ./frontend/index.html:/app/index.html
+      - ./frontend/src/styles:/app/src/styles
+    environment:
+      - NODE_ENV=dev
     restart: always
-    links:
+    ports:
+      - 8081:8081
+    depends_on:
+      - backend
+    networks:
       - backend
 
   frontend:
     image: nginx:stable-alpine
+    container_name: pwndoc-ng-frontend-proxy
     user: root
     volumes:
       - ./frontend/.docker/nginx.dev.conf:/etc/nginx/conf.d/default.conf
@@ -61,9 +81,25 @@ services:
     restart: always
     ports:
       - 8443:8443
+    depends_on:
+      - frontend-app
+      - backend
     links:
       - frontend-app
+      - backend
+    networks:
+      - backend
+
+  languagetool:
+    image: erikvl87/languagetool
+    container_name: pwndoc-ng-languagetool-dev
+    restart: always
+    networks:
       - backend
 
 volumes:
   mongo-data-dev:
+
+networks:
+  backend:
+    driver: bridge
