+++ docker-compose.yml
@@ -24,6 +24,7 @@ services:
       - mongodb
     environment:
       - COLLAB_WEBSOCKET_PORT=8440
+      - NODE_ENV=production
     restart: always
     ports:
       - 4242:4242
@@ -37,13 +38,13 @@ services:
     image: pwndoc-ng/pwndoc-ng:frontend
     container_name: pwndoc-ng-frontend
     restart: always
-    environment:
-      - Java_Xms=512m
-      - Java_Xmx=2g
     ports:
       - 8443:8443
+    depends_on:
+      - backend
     networks:
       - backend
+
   languagetool:
     image: erikvl87/languagetool
     container_name: pwndoc-ng-languagetool
