+++ frontend/.docker/nginx.dev.conf
@@ -3,45 +3,75 @@ server {
     server_name         localhost;
     ssl_certificate     /etc/nginx/ssl/server.cert;
     ssl_certificate_key /etc/nginx/ssl/server.key;
-    ssl_protocols       TLSv1 TLSv1.1 TLSv1.2;
+    ssl_protocols       TLSv1.2 TLSv1.3;
     ssl_ciphers         HIGH:!aNULL:!MD5;
 
-    #charset koi8-r;
-    #access_log  /var/log/nginx/host.access.log  main;
-
+    # Proxy to Vite development server
     location / {
         proxy_pass  https://frontend-app:8081;
+        proxy_ssl_verify off;
         proxy_read_timeout 3600;
+        proxy_set_header Host $host;
+        proxy_set_header X-Real-IP $remote_addr;
+        proxy_set_header X-Forwarded-For $proxy_add_x_forwarded_for;
+        proxy_set_header X-Forwarded-Proto $scheme;
+        
+        # WebSocket support for Vite HMR
+        proxy_http_version 1.1;
+        proxy_set_header Upgrade $http_upgrade;
+        proxy_set_header Connection "upgrade";
     }
 
+    # API proxy to backend
     location /api {
-        proxy_pass  https://backend:5252;
+        proxy_pass  http://backend:5252;
         proxy_read_timeout 3600;
+        proxy_set_header Host $host;
+        proxy_set_header X-Real-IP $remote_addr;
+        proxy_set_header X-Forwarded-For $proxy_add_x_forwarded_for;
+        proxy_set_header X-Forwarded-Proto $scheme;
     }
 
+    # LanguageTool proxy
+    location /v2 {
+        proxy_pass  http://pwndoc-ng-languagetool-dev:8010;
+        proxy_read_timeout 3600;
+        proxy_set_header Host $host;
+        proxy_set_header X-Real-IP $remote_addr;
+        proxy_set_header X-Forwarded-For $proxy_add_x_forwarded_for;
+        proxy_set_header X-Forwarded-Proto $scheme;
+    }
+
+    # Socket.io proxy
     location /socket.io {
-        proxy_pass  https://backend:5252;
+        proxy_pass  http://backend:5252;
         proxy_http_version 1.1;
         proxy_set_header Upgrade $http_upgrade;
         proxy_set_header Connection "upgrade";
         proxy_set_header Host $host;
+        proxy_set_header X-Real-IP $remote_addr;
+        proxy_set_header X-Forwarded-For $proxy_add_x_forwarded_for;
+        proxy_set_header X-Forwarded-Proto $scheme;
     }
+
+    # Collaboration WebSocket proxy
     location /collab {
         proxy_pass  http://backend:8440;
         proxy_http_version 1.1;
         proxy_set_header Upgrade $http_upgrade;
         proxy_set_header Connection "upgrade";
         proxy_set_header Host $host;
+        proxy_set_header X-Real-IP $remote_addr;
+        proxy_set_header X-Forwarded-For $proxy_add_x_forwarded_for;
+        proxy_set_header X-Forwarded-Proto $scheme;
     }
 
-    #error_page  404              /404.html;
-
-    # redirect server error pages to the static page /50x.html
-    #
+    # Error pages
     error_page   500 502 503 504  /50x.html;
     location = /50x.html {
         root   /usr/share/nginx/html;
     }
 
+    # Increase client body size for file uploads
     client_max_body_size 10M;
 }
