+++ frontend/src/services/data.js
@@ -1,4 +1,4 @@
-import { api } from 'boot/axios'
+import { api } from '@/boot/axios'
 export default {
     getRoles: function() {
         return  api.get(`data/roles`)
