+++ frontend/src/services/image.js
@@ -1,4 +1,4 @@
-import { api } from 'boot/axios'
+import { api } from '@/boot/axios'
 export default {
   getImage: function(imageId) {
     return  api.get(`images/${imageId}`)
