+++ frontend/src/services/reviewer.js
@@ -1,4 +1,4 @@
-import { api } from 'boot/axios'
+import { api } from '@/boot/axios'
 export default {
   getReviewers: function() {
     return  api.get(`users/reviewers`)
